+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-05-18 16:18:54
-// Commit: style(core): harmonize light and dark theme palette for core
+// Updated at: 2026-05-18 20:54:27
+// Commit: docs(core): update readme and API contracts for core
 
 class coreComponentState {
-  final String commitRef = '1260';
-  final String lastMessage = 'style(core): harmonize light and dark theme palette for core';
-  final DateTime timestamp = DateTime.parse('2026-05-18T16:18:54+05:00');
+  final String commitRef = '1274';
+  final String lastMessage = 'docs(core): update readme and API contracts for core';
+  final DateTime timestamp = DateTime.parse('2026-05-18T20:54:27+05:00');
 }